Clean up and comment Translate app
