+++ examples/ios/swift/Translate/Language.swift
@@ -1,12 +1,6 @@
 import Foundation
 import RealmSwift
 
-enum LanguageDirection: Int, Codable {
-    case unknown
-    case leftToRight
-    case rightToLeft
-}
-
 extension Locale.LanguageDirection: Codable {
 }
 
@@ -14,6 +8,7 @@ enum LanguageState {
     case from, to
 }
 
+/// A spoken language
 class Language: Object, Codable {
     static let defaultPrimary: Language = {
         let language = Language()
@@ -38,10 +33,13 @@ class Language: Object, Codable {
         return language
     }()
 
+    /// The name of the language
     @objc dynamic var name: String = ""
+    /// The name of the language in the language itself
     @objc dynamic var nativeName: String = ""
+    /// The ISO language code
     @objc dynamic var languageCode: String = ""
-
+    /// The direction the characters of the language are written in
     lazy var direction: Locale.LanguageDirection =
         Locale.characterDirection(forLanguage: languageCode)
