+++ examples/ios/swift/Translate/SceneDelegate.swift
@@ -1,27 +1,6 @@
 import UIKit
 import SwiftUI
 import RealmSwift
-import os
-
-class Translation: Object, Codable {
-    private enum CodingKeys: String, CodingKey {
-        case originalText, translatedText, to, from
-    }
-
-    @objc dynamic var originalText: String
-    @objc dynamic var translatedText: String
-    @objc dynamic var to: String
-    @objc dynamic var from: String
-
-    @objc dynamic var isSaved: Bool = false
-}
-
-class UserSettings: Object {
-    @objc dynamic var lastLangFrom: String = Locale.current.identifier
-    @objc dynamic var lastLangTo: String = "fr"
-
-    let translations: RealmSwift.List<Translation> = RealmSwift.List()
-}
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
@@ -40,20 +19,5 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             self.window = window
             window.makeKeyAndVisible()
         }
-    }
-
-    func sceneDidDisconnect(_ scene: UIScene) {
-    }
-
-    func sceneDidBecomeActive(_ scene: UIScene) {
-    }
-
-    func sceneWillResignActive(_ scene: UIScene) {
-    }
-
-    func sceneWillEnterForeground(_ scene: UIScene) {
-    }
-
-    func sceneDidEnterBackground(_ scene: UIScene) {
     }
 }
